Incoporate PR review feedback
